% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Arbeit}
\alias{Arbeit}
\title{Befragung zur Arbeitszufriedenheit}
\format{
176 Fälle bei 10 Variablen (data.frame)
\describe{
\item{S1}{Mein Arbeitgeber achtet auf einen ordnungsgemäßen Arbeitsplatz.}
\item{S2}{Mein Arbeitgeber bietet mir die modernste / aktuellste Sicherheitsausstattung.}
\item{S3}{Mein Arbeitgeber hat Arbeitssicherheit Vorrang vor der Produktion.}
\item{Q1}{In meinem Betrieb geht Produktqualität vor Quantität.}
\item{Q2}{In meinem Betrieb wird Ressourcen Verschwendung vermieden.}
\item{Q3}{In meinem Betrieb versuchen erkrankte Mitarbeiter so schnell wie möglich an den Arbeitsplatz zurückzukehren.}
\item{Z1}{Ich bin zufrieden mit dem Organisationsklima und dem Teamgeist.}
\item{Z2}{Ich bin zufrieden mit den Arbeitsbedingungen.}
\item{Z3}{Ich bin zufrieden mit den Karrieremöglichkeiten in meinem Betrieb.}
\item{Lohn}{Wohlbefinden}
}
}
\usage{
Arbeit
}
\description{
176 Arbeitskräfte beantworten 10 Items zur Arbeitssituation.
}
\keyword{datasets}
