% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{BTS}
\alias{BTS}
\title{Wrapper for main BTS functionality.}
\usage{
BTS(in_path, out_path, d = 2)
}
\arguments{
\item{in_path}{path containing input data, in individual
directories for each locus.}

\item{out_path}{path for writing output.}

\item{d}{Positive integer, maximum size of causal sets
considered.}
}
\description{
Read input, compute and write output.
}
