% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_utils.R
\name{compute_likelihoods}
\alias{compute_likelihoods}
\title{Enumerate configurations; compute and store likelihoods.}
\usage{
compute_likelihoods(in_path, ld_files, loci, d = 2, thresh = 12)
}
\arguments{
\item{in_path}{A path.}

\item{ld_files}{A vector of files.}

\item{loci}{A vector of locus names.}

\item{d}{Maximal number of independent signals considered.}

\item{thresh}{Numeric. Configurations are not stored if their log 
Bayes factor is smaller than that of the size-1 configuration with 
largest Bayes factor, minus the threshold.}
}
\description{
Enumerate over all configurations with at most d
causal signals. Compute log likelihoods and store them for future
use in posterior computations.
}
