% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Depression}
\alias{Depression}
\title{Depression}
\format{
A data frame with 40 rows and 3 variables:
\describe{
\item{Depr}{Depression (d = depression; nd = no depression)}
\item{Gew}{weight in kg}
\item{Konz}{Concentration Test }
}
}
\usage{
Depression
}
\description{
A fictitious dataset for depression classification.
}
\keyword{datasets}
