% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FW1}
\alias{FW1}
\title{Fehlende Werte 1 (Missing Values 1)}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{IQ}{IQ}
\item{Alter}{Age}
\item{LZ}{life satisfaction}
\item{KFZ}{car}
}
}
\usage{
FW1
}
\description{
A fictitious dataset containing 4 variables with 30 missing values.
}
\keyword{datasets}
