% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FW_abbrecher}
\alias{FW_abbrecher}
\title{Fehlende Werte 3 (Missing Values 3)}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{IQ}{IQ}
\item{Alter}{Age}
\item{LZ}{life satisfaction}
\item{KFZ}{car}
}
}
\usage{
FW_abbrecher
}
\description{
A fictitious dataset containing 4 variables and a typical pattern of missing values which indicates a break off.
}
\keyword{datasets}
