% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FW_mar}
\alias{FW_mar}
\title{Fehlende Werte 2 (Missing Values 2)}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{IQ}{IQ}
\item{Alter}{Age}
\item{LZ}{life satisfaction}
\item{KFZ}{car}
}
}
\usage{
FW_mar
}
\description{
A fictitious dataset containing 4 variables with 45 values missing at random.
}
\keyword{datasets}
