% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FW_mcar}
\alias{FW_mcar}
\title{Fehlende Werte 1 (Missing Values 1)}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{IQ}{IQ}
\item{Alter}{Alter}
\item{LZ}{Lebenszufriedenheit}
\item{KFZ}{Besitzen Sie ein Auto?}
}
}
\usage{
FW_mcar
}
\description{
A fictitious dataset containing 4 variables with 30 values missing completly at random.
}
\keyword{datasets}
