% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{get_annot_inputs}
\alias{get_annot_inputs}
\title{Reads and organizes functional annotations.}
\usage{
get_annot_inputs(path, loci)
}
\arguments{
\item{path}{Path to annotations.}

\item{loci}{A vector of loci.}
}
\description{
Loci must have corresponding directories
in the provided path. Assuming each locus directory
contains annotation files in the format
annotations_[group].tsv. [group] could be a tissue
category, e.g. Brain or Blood.
}
