% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_utils.R
\name{get_model_summary}
\alias{get_model_summary}
\title{Summary and comparison of many models.}
\usage{
get_model_summary(results, annot_inputs)
}
\arguments{
\item{results}{List of results for models.}

\item{annot_inputs}{List of annotations for all models and loci.}
}
\description{
Returns a data frame of models, with enrichment
parameters, log likelihood, number of annotated variants, sorted
by p-value.
}
