% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdp}
\alias{hdp}
\title{Hospital Doctor Patient Dataset}
\format{
A data frame with 8525 rows and 4 variables:
\describe{
\item{remission}{cancer in remission? (TRUE/FALSE)}
\item{CancerStage}{stage of cancer}
\item{Experience}{experience of doctor}
\item{DID}{doctor ID}
}
}
\source{
\url{https://stats.idre.ucla.edu/r/codefragments/mesimulation/}
}
\usage{
hdp
}
\description{
A dataset containing simulated data on lung cancer in a
three-level-hierarchical structure
}
\keyword{datasets}
