% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icecream}
\alias{icecream}
\title{Ice cream rankings}
\format{
A data frame with 150 rows and 7 variables:
\describe{
\item{Observations}{profile number}
\item{Flavor}{flavor of ice cream}
\item{Packaging}{packaging of ice cream}
\item{Light}{fat content of ice cream}
\item{Organic}{is ice cream organic?}
\item{individual}{ID for participant}
\item{ranking}{ranking of ice cream, from 1 to 10}
}
}
\source{
\url{https://help.xlstat.com/s/article/conjoint-analysis-in-excel-tutorial-new?language=en_US}
\url{http://users.telenet.be/samuelfranssens/tutorial_data/icecream.xlsx}
}
\usage{
icecream
}
\description{
A dataset containing ice cream attributes and rankings from 15 (fictitious)
participants.
}
\keyword{datasets}
