% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lehr_Lern}
\alias{Lehr_Lern}
\title{Mediation von Lernleistung (Meadiation of learning achievement)}
\format{
A data frame with 30 rows and 3 variables:
\describe{
\item{Motivation}{Motivation}
\item{Lernleistung}{Learning achievement}
\item{Unterrichtsguete}{Teaching quality}
}
}
\usage{
Lehr_Lern
}
\description{
A fictitious dataset containing 3 variables for mediation
}
\keyword{datasets}
