% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mehrere_AVs}
\alias{Mehrere_AVs}
\title{Mehrere AVs in einem Einzelfall}
\format{
Erwins Daten aus einem AB-Einzelfalldesign
\describe{
\item{Fall}{Name der ProbandInnen}
\item{Phase}{Phase im Einzelfalldesign}
\item{Pos_Emot}{Positive Emotionen}
\item{Neg_Emot}{Negative Emotionen}
\item{Angst}{Angstrating}
\item{Depression}{Depressionrating}
}
}
\usage{
Mehrere_AVs
}
\description{
Mehrere AVs in einem Einzelfall
}
\keyword{datasets}
