% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Obst}
\alias{Obst}
\title{Obst}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{Apfel}{Rating of apple}
\item{Birne}{Rating of pear}
\item{Kirsche}{Rating of cherry}
\item{Aprikose}{Rating of apricot}
}
}
\usage{
Obst
}
\description{
A fictitious data set with preferences for four different fruits.
}
\keyword{datasets}
