% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedersen_2002}
\alias{pedersen_2002}
\title{pedersen_2002}
\format{
A dataset of 206 persons and their desired number of sexual partners.
A modified Version of Pedersen et al. (2002).
\describe{
\item{person_id}{ID of participants}
\item{desired}{Number of desired sexual partners}
\item{sex}{male vs. female}
}
}
\usage{
pedersen_2002
}
\description{
Number of desired sexual partners
}
\keyword{datasets}
