% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Power_Pose}
\alias{Power_Pose}
\title{Power_Pose}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{Pose}{Experimental condition: high powerpose vs. low powerpose}
\item{Dominanz}{Dominance rating by the neutral obeserver}
\item{Aktivitaet}{Activity}
\item{Fuehrung}{External rating of leadership}
\item{Moderator}{Moderation effect of Dom x Pose}
}
}
\usage{
Power_Pose
}
\description{
A dataset of 16 cases beeing part of an powerposing experiment.
}
\keyword{datasets}
