% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pub_bias0}
\alias{pub_bias0}
\title{pub_bias0}
\format{
A data frame with 56 rows and 3 variables:
\describe{
\item{N}{Number of participants}
\item{r}{Effectsize r}
\item{p}{p-value (two.tailed)}
}
}
\usage{
pub_bias0
}
\description{
A fictitious dataset with effectsizes from 56 studies.
}
\keyword{datasets}
