% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pub_bias1}
\alias{pub_bias1}
\title{A dataset with publication bias}
\format{
A data frame with 28 rows and 3 variables:
\describe{
\item{N}{Number of participants}
\item{r}{Effectsize r}
\item{p}{p-value (two.tailed)}
}
}
\usage{
pub_bias1
}
\description{
A dataset generated fram pub_bias0 with no nonsignificant p-values.
}
\keyword{datasets}
