% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_utils.R
\name{run_em}
\alias{run_em}
\title{Fit a model using EM algorithm.}
\usage{
run_em(configs, annot, max_iter = 20, tol = 0.01)
}
\arguments{
\item{configs}{Configurations of causal variants and their
log Bayes factors.}

\item{annot}{For each locus, a matrix of annotations.}

\item{max_iter}{Positive integer, maximum number of 
EM iterations after which algorithm stops regardless of
convergence.}

\item{tol}{Tolerance for convergence.}
}
\value{
A list of items:
\itemize{
\item posteriors - A list giving the vector of variant posterior
probabilities for all loci.
\item enrich - A named vector giving the values of track enrichments.
\item log_post - The total data likelihood under this model.
\item log_post_locus - A vector of data likelihoods for each locus.
}
}
\description{
Given likelihoods and annotations for all loci,
find optimal values for variant posterior probabilities to
be causal, as well as enrichment of annotations in causal
variants. Uses Expectation-Maximization to optimize total
data likelihood.
}
