% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school}
\alias{school}
\title{School data set}
\format{
A data frame with 1200 rows and 7 variables:
\describe{
\item{id}{student ID}
\item{extro}{probably extroversion}
\item{open}{probably openness for experience}
\item{agree}{probably agreeableness}
\item{social}{?}
\item{class}{ID for class}
\item{school}{ID for school}
}
}
\source{
\url{http://bayes.acs.unt.edu:8083/BayesContent/class/Jon/R_SC/Module9/LMM_Examples.R}
\url{http://bayes.acs.unt.edu:8083/BayesContent/class/Jon/R_SC/Module9/lmm.data.txt}
}
\usage{
school
}
\description{
A dataset containing simulated data on students in different schools. Note
that the variable names are not important as this is simulated data.
}
\keyword{datasets}
