% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sedlmeier}
\alias{Sedlmeier}
\title{Sedlmeier}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{Studie}{Author of the study}
\item{Effekt_r}{Effectsize r}
\item{N}{Number of participants}
\item{AV}{Measurrement of the dependent variable}
}
}
\usage{
Sedlmeier
}
\description{
A shortened dataset from Sedlmeier et al. (2018). The effect of six
Meditation studies.
}
\keyword{datasets}
