% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{St_dist_aggr}
\alias{St_dist_aggr}
\title{Straftaten agreggierte Distanzen}
\format{
A distance matrix with 10 objects:
\describe{
\item{Unt}{failure to assist a person in danger}
\item{Ste}{tax evasion}
\item{Sac}{damage to property}
\item{Rau}{misappropriation}
\item{Wid}{resistance against enforcement officers}
\item{Tru}{drunk driving}
\item{Sac}{damage to property}
\item{Ein}{burglary}
\item{Ver}{rape}
\item{Koe}{bodily injury}
}
}
\usage{
St_dist_aggr
}
\description{
A distance matrix with mean aggregated distances for ten criminal offenses
}
\keyword{datasets}
