% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Stadt_dist}
\alias{Stadt_dist}
\title{Distanzen von 5 deutschen Städten}
\format{
A symmetric data.frame with 5 rows.
}
\usage{
Stadt_dist
}
\description{
Stadt_dist ist eine Matrix mit den Entfernungen (Luftlinie in km) zwischen
5 deutschen Städten.
}
\keyword{datasets}
