% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Strafen}
\alias{Strafen}
\title{Strafen}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{Apfel}{Rating of apple}
\item{Birne}{Rating of pear}
\item{Kirsche}{Rating of cherry}
\item{Aprikose}{Rating of apricot}
}
}
\usage{
Strafen
}
\description{
EMPTY ....
}
\keyword{datasets}
