% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Straftaten}
\alias{Straftaten}
\title{Straftaten}
\format{
A data frame with 30 rows and 45 variables:
\describe{
\item{UntSac}{similaryt between failure to assist a person in danger and damage to property}
\item{SteTru}{similaryt between misaptax evasionpropriation and drunk driving}
\item{SacWid}{similaryt between damage to property and resistance against enforcement officers}
\item{RauSte}{similaryt between misappropriation and tax evasion}
\item{SteUnt}{similaryt between tax evasion and failure to assist a person in danger}
\item{WidKoe}{similaryt between resistance against enforcement officers and damage to property}

\item{SteWid}{similaryt between tax evasion and resistance against enforcement officers}
\item{EinTru}{similaryt between misappropriation and drunk driving}
\item{KoeSac}{similaryt between misappropriation and damage to property}
\item{SteEin}{similaryt between tax evasion and damage to property}
\item{EinKoe}{similaryt between misappropriation and damage to property}
\item{VerWid}{similaryt between misappropriation and resistance against enforcement officers}
\item{WidRau}{similaryt between resistance against enforcement officers and damage to property}

\item{UntWid}{similaryt between failure to assist a person in danger and resistance against enforcement officers}
\item{KoeTru}{similaryt between misappropriation and drunk driving}
\item{EinUnt}{similaryt between misappropriation and failure to assist a person in danger}
\item{TruSac}{similaryt between drunk driving and damage to property}
\item{UntVer}{similaryt between failure to assist a person in danger and damage to property}
\item{TruUnt}{similaryt between drunk driving and failure to assist a person in danger}

\item{WahSac}{similaryt between misappropriation and damage to property}
\item{WidTru}{similaryt between resistance against enforcement officers and drunk driving}
\item{TruVer}{similaryt between drunk driving and damage to property}
\item{KoeSte}{similaryt between misappropriation and tax evasion}
\item{RauUnt}{similaryt between misappropriation and failure to assist a person in danger}
\item{KoeRau}{similaryt between misappropriation and damage to property}
\item{WidEin}{similaryt between resistance against enforcement officers and damage to property}

\item{SteWah}{similaryt between tax evasion and damage to property}
\item{KoeVer}{similaryt between misappropriation and damage to property}
\item{VerWah}{similaryt between misappropriation and damage to property}
\item{EinWah}{similaryt between misappropriation and damage to property}
\item{RauEin}{similaryt between misappropriation and damage to property}
\item{UntWah}{similaryt between failure to assist a person in danger and damage to property}

\item{SacVer}{similaryt between damage to property and damage to property}
\item{RauWah}{similaryt between misappropriation and damage to property}
\item{WahKoe}{similaryt between misappropriation and damage to property}
\item{RauTru}{similaryt between misappropriation and drunk driving}
\item{SteVer}{similaryt between tax evasion and damage to property}
\item{SacEin}{similaryt between damage to property and damage to property}
\item{RauVer}{similaryt between misappropriation and damage to property}

\item{WahWid}{similaryt between misappropriation and resistance against enforcement officers}
\item{RauSac}{similaryt between misappropriation and damage to property}
\item{TruWah}{similaryt between drunk driving and damage to property}
\item{KoeUnt}{similaryt between misappropriation and failure to assist a person in danger}
\item{VerEin}{similaryt between misappropriation and damage to property}
\item{SteSac}{similaryt between tax evasion and damage to property}
}
}
\usage{
Straftaten
}
\description{
A data set with similarity ratings for ten criminal offenses measured by pairwise comparison.
}
\keyword{datasets}
