% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VL17}
\alias{VL17}
\title{VL17}
\format{
A dataset of 50 preference ratings for cats and dogs.
\describe{
\item{Hunde}{Preference for dogs}
\item{desired}{Preference for cats}
}
}
\usage{
VL17
}
\description{
Preference for cats and dogs
}
\keyword{datasets}
