% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wellbeing_Daten}
\alias{Wellbeing_Daten}
\title{6 Meditierende im Multiple Baseline Design (scan-Paket)}
\format{
6 TeilnehmerInnen aus dem Multiple Baseline Experiment von Matko
et al. 2021 zur Wirkung von Meditation als scan-Objekt
\describe{
\item{baseline}{Dauer der Baseline (d)}
\item{study_day}{Studientag}
\item{wellbeing}{Wohlbefinden}
\item{phase}{Phase im Multiple Baseline Design}
}
}
\usage{
Wellbeing_Daten
}
\description{
6 Meditierende im Multiple Baseline Design (scan-Paket)
}
\keyword{datasets}
