% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wellbeing_kurz}
\alias{Wellbeing_kurz}
\title{Lebenszufriedenheit von 6 Einzelfällen}
\format{
6 TeilnehmerInnen eines multiple Baseline Experiment von Matko
et al. 2021 zur Wirkung von Meditation als data.frame im long-Format.
\describe{
\item{baseline}{Dauer der Baseline (d)}
\item{study_day}{Studientag}
\item{wellbeing}{Wohlbefinden}
\item{phase}{Phase im Multiple Baseline Design}
}
}
\usage{
Wellbeing_kurz
}
\description{
6 Meditierende im Multiple Baseline Design (data.frame)
}
\keyword{datasets}
