% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_utils.R
\name{compute_configs}
\alias{compute_configs}
\title{Enumerate configurations; compute and store likelihoods.}
\usage{
compute_configs(path, loci, d = 2, thresh = 12)
}
\arguments{
\item{path}{A path to data.}

\item{loci}{A vector of locus names; each locus must have a sub-directory
under the given path.}

\item{d}{Integer, maximal number of independent signals considered.}

\item{thresh}{Numeric. Configurations are not stored if their log 
Bayes factor is smaller than that of the size 1 configuration with 
largest Bayes factor, minus the threshold.}
}
\value{
A list of two elements: an integer matrix of shape (n_config,d)
and a numeric vector of length n_config, where n_config is the number of 
configurations of d variants with large enough Bayes factor, at most
choose(n_variants, d). The rows of the matrix are indices of variants
giving a causal configuration, and the corresponding row of the vector is
the log Bayes factor of the configuration.
}
\description{
Enumerate over all configurations with at most d
causal signals. Compute log likelihoods and store them for future
use in posterior computations.
}
