% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{get_annot_inputs}
\alias{get_annot_inputs}
\title{Reads and organizes functional annotations.}
\usage{
get_annot_inputs(in_path, loci)
}
\arguments{
\item{in_path}{Path to annotations.}

\item{loci}{A vector of loci.}
}
\description{
Currently assumes annotations are organized by tissue.
}
