% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Stadt_dist}
\alias{Stadt_dist}
\title{Distances of 5 german cities}
\format{
A symmetric data.frame with 5 rows.
}
\usage{
Stadt_dist
}
\description{
Stdt_dist gives the distances (in km) between 5 cities in Germany.
}
\keyword{datasets}
